Optimizing FS's a little bit

+++ src/plugins/PluginShapes.h
@@ -107,7 +107,7 @@ struct alignas(16) ShapeData2D { // 160 bytes
 
     float opacity;
 
-    ivec4 squareInitialCoords;  // 16 bytes (X, Y, Width, Height)
+    vec4 squareInitialCoords;   // 16 bytes (X, Y, Width, Height)
     vec4 squareFinalCoords;     // 16 bytes (X, Y, Width, Height)
 
     vec4 fadeBorderSize;        // 16 bytes (left fade, top fade, right fade, down fade)
@@ -128,10 +128,10 @@ struct alignas(16) ShapeData2D { // 160 bytes
 
         opacity = opacity * sourcePercentage + finalShape.opacity * finalPercentage;
 
-        squareInitialCoords.x = (int)(squareInitialCoords.x * sourcePercentage + finalShape.squareInitialCoords.x * finalPercentage);
-        squareInitialCoords.y = (int)(squareInitialCoords.y * sourcePercentage + finalShape.squareInitialCoords.y * finalPercentage);
-        squareInitialCoords.z = (int)(squareInitialCoords.z * sourcePercentage + finalShape.squareInitialCoords.z * finalPercentage);
-        squareInitialCoords.w = (int)(squareInitialCoords.w * sourcePercentage + finalShape.squareInitialCoords.w * finalPercentage);
+        squareInitialCoords.x = squareInitialCoords.x * sourcePercentage + finalShape.squareInitialCoords.x * finalPercentage;
+        squareInitialCoords.y = squareInitialCoords.y * sourcePercentage + finalShape.squareInitialCoords.y * finalPercentage;
+        squareInitialCoords.z = squareInitialCoords.z * sourcePercentage + finalShape.squareInitialCoords.z * finalPercentage;
+        squareInitialCoords.w = squareInitialCoords.w * sourcePercentage + finalShape.squareInitialCoords.w * finalPercentage;
 
         squareFinalCoords.x = squareFinalCoords.x * sourcePercentage + finalShape.squareFinalCoords.x * finalPercentage;
         squareFinalCoords.y = squareFinalCoords.y * sourcePercentage + finalShape.squareFinalCoords.y * finalPercentage;
@@ -157,7 +157,7 @@ struct alignas(16) ShapeData3D {
     int corner;     // 4 bytes
     float hudScale; // 4 bytes
 
-    ivec4 squareInitialCoords; // 16 bytes (X, Y, Width, Height)
+    vec4 squareInitialCoords; // 16 bytes (X, Y, Width, Height)
 
     vec4 margin; // 16 bytes (left, top, right, bottom)
 
@@ -374,10 +374,10 @@ public:
         shapeBuilder._shape = shape_Square;
         shapeBuilder.shapeData.sourceScale.x = 1.0;
         shapeBuilder.shapeData.sourceScale.y = 1.0;
-        shapeBuilder.shapeData.squareInitialCoords.x = 0;
-        shapeBuilder.shapeData.squareInitialCoords.y = 0;
-        shapeBuilder.shapeData.squareInitialCoords.z = 256;
-        shapeBuilder.shapeData.squareInitialCoords.w = 192;
+        shapeBuilder.shapeData.squareInitialCoords.x = 0.0;
+        shapeBuilder.shapeData.squareInitialCoords.y = 0.0;
+        shapeBuilder.shapeData.squareInitialCoords.z = 256.0;
+        shapeBuilder.shapeData.squareInitialCoords.w = 192.0;
         shapeBuilder.shapeData.opacity = 1.0;
         shapeBuilder._hudScale = 1.0;
         shapeBuilder._colorIndex = 0;
@@ -420,15 +420,15 @@ public:
     }
     ShapeBuilder2D& fromPosition(int x, int y) {
         if (_shape == shape_Square) {
-            shapeData.squareInitialCoords.x = x;
-            shapeData.squareInitialCoords.y = y;
+            shapeData.squareInitialCoords.x = (float)x;
+            shapeData.squareInitialCoords.y = (float)y;
         }
         return *this;
     }
     ShapeBuilder2D& withSize(int width, int height) {
         if (_shape == shape_Square) {
-            shapeData.squareInitialCoords.z = width;
-            shapeData.squareInitialCoords.w = height;
+            shapeData.squareInitialCoords.z = (float)width;
+            shapeData.squareInitialCoords.w = (float)height;
         }
         return *this;
     }
@@ -617,7 +617,7 @@ public:
 
     ShapeData2D build(float aspectRatio) {
         if (_fromBottomScreen) {
-            shapeData.squareInitialCoords.y += 192;
+            shapeData.squareInitialCoords.y += 192.0;
         }
 
         if (_shape == shape_Square) {
@@ -645,10 +645,10 @@ public:
         shapeBuilder._shape = shape_Square;
         shapeBuilder.shapeData.sourceScale.x = 1.0;
         shapeBuilder.shapeData.sourceScale.y = 1.0;
-        shapeBuilder.shapeData.squareInitialCoords.x = 0;
-        shapeBuilder.shapeData.squareInitialCoords.y = 0;
-        shapeBuilder.shapeData.squareInitialCoords.z = 256;
-        shapeBuilder.shapeData.squareInitialCoords.w = 192;
+        shapeBuilder.shapeData.squareInitialCoords.x = 0.0;
+        shapeBuilder.shapeData.squareInitialCoords.y = 0.0;
+        shapeBuilder.shapeData.squareInitialCoords.z = 256.0;
+        shapeBuilder.shapeData.squareInitialCoords.w = 192.0;
         shapeBuilder.shapeData.hudScale = SCREEN_SCALE;
         shapeBuilder.shapeData.corner = corner_PreservePosition;
         shapeBuilder.shapeData.zRange.x = -1.0;
@@ -732,25 +732,25 @@ public:
     }
     ShapeBuilder3D& fromPosition(int x, int y) {
         if (_shape == shape_Square) {
-            shapeData.squareInitialCoords.x = x;
-            shapeData.squareInitialCoords.y = y;
+            shapeData.squareInitialCoords.x = (float)x;
+            shapeData.squareInitialCoords.y = (float)y;
         }
         return *this;
     }
     ShapeBuilder3D& withSize(int width, int height) {
         if (_shape == shape_Square) {
-            shapeData.squareInitialCoords.z = width;
-            shapeData.squareInitialCoords.w = height;
+            shapeData.squareInitialCoords.z = (float)width;
+            shapeData.squareInitialCoords.w = (float)height;
         }
         return *this;
     }
     ShapeBuilder3D& includeOutOfBoundsPolygons() {
         if (_shape == shape_Square) {
-            int margin = 20;
+            float margin = 20;
             shapeData.squareInitialCoords.x = -margin;
             shapeData.squareInitialCoords.y = -margin;
-            shapeData.squareInitialCoords.z = 256 + margin*2;
-            shapeData.squareInitialCoords.w = 192 + margin*2;
+            shapeData.squareInitialCoords.z = 256.0 + margin*2;
+            shapeData.squareInitialCoords.w = 192.0 + margin*2;
         }
         return *this;
     }
